Rename auth file -> useAuth
