--- ui/src/services/auth/index.js
@@ -1,4 +1,0 @@
-import auth from './auth';
-
-export default auth;
-
