+++ ui/src/services/useAuth/index.js
@@ -1,0 +1,4 @@
+import useAuth from './useAuth';
+
+export default useAuth;
+
